requirements.txt
Flask~=2.2.2
PyJWT~=2.6.0
pymongo~=4.3.2
bcrypt~=4.0.1
bson~=0.5.10